fix: missed refactoring get responses somehow
